Remove EN permissions checks on Notify Others screen in the production flavor
