+++ app/src/calibration/java/org/covidwatch/android/ui/reporting/NotifyOthers.kt
@@ -12,6 +12,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.viewModelScope
 import com.google.android.gms.nearby.exposurenotification.TemporaryExposureKey
+import kotlinx.coroutines.launch
 import org.covidwatch.android.BuildConfig
 import org.covidwatch.android.R
 import org.covidwatch.android.data.PositiveDiagnosisReport
@@ -20,6 +21,8 @@ import org.covidwatch.android.databinding.DialogRiskLevelsBinding
 import org.covidwatch.android.domain.ExportDiagnosisKeysAsFileUseCase
 import org.covidwatch.android.domain.StartUploadDiagnosisKeysWorkUseCase
 import org.covidwatch.android.exposurenotification.ExposureNotificationManager
+import org.covidwatch.android.exposurenotification.ExposureNotificationManager.Companion.PERMISSION_KEYS_REQUEST_CODE
+import org.covidwatch.android.exposurenotification.ExposureNotificationManager.Companion.PERMISSION_START_REQUEST_CODE
 import org.covidwatch.android.extension.launchUseCase
 import org.covidwatch.android.extension.observeEvent
 import org.covidwatch.android.extension.send
@@ -108,10 +111,9 @@ class NotifyOthersFragment : BaseNotifyOthersFragment() {
 class NotifyOthersViewModel(
     private val exportDiagnosisKeysAsFileUseCase: ExportDiagnosisKeysAsFileUseCase,
     private val startUploadDiagnosisKeysWorkUseCase: StartUploadDiagnosisKeysWorkUseCase,
-    enManager: ExposureNotificationManager,
+    private val enManager: ExposureNotificationManager,
     positiveDiagnosisRepository: PositiveDiagnosisRepository
 ) : BaseNotifyOthersViewModel(
-    enManager,
     positiveDiagnosisRepository
 ) {
     private val defaultRiskLevel = 6
@@ -140,7 +142,33 @@ class NotifyOthersViewModel(
         _chooseShareMethod.send()
     }
 
-    override fun onTekHistory(teks: MutableList<TemporaryExposureKey>) {
+    override fun sharePositiveDiagnosis() {
+        viewModelScope.launch {
+            enManager.isEnabled().success { enabled ->
+                if (enabled) {
+                    shareReport()
+                } else {
+                    withPermission(PERMISSION_START_REQUEST_CODE) {
+                        enManager.start().apply {
+                            success { shareReport() }
+                            failure { handleStatus(it) }
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    private suspend fun shareReport() {
+        withPermission(PERMISSION_KEYS_REQUEST_CODE) {
+            enManager.temporaryExposureKeyHistory().apply {
+                success { onTekHistory(it) }
+                failure { handleStatus(it) }
+            }
+        }
+    }
+
+    private fun onTekHistory(teks: MutableList<TemporaryExposureKey>) {
         _setTransmissionLevelRisk.send(teks.map { defaultRiskLevel })
     }
