+++ app/src/main/java/org/covidwatch/android/ui/reporting/BaseNotifyOthersViewModel.kt
@@ -3,19 +3,12 @@ package org.covidwatch.android.ui.reporting
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.map
-import androidx.lifecycle.viewModelScope
-import com.google.android.gms.nearby.exposurenotification.TemporaryExposureKey
-import kotlinx.coroutines.launch
 import org.covidwatch.android.data.positivediagnosis.PositiveDiagnosisRepository
-import org.covidwatch.android.exposurenotification.ExposureNotificationManager
-import org.covidwatch.android.exposurenotification.ExposureNotificationManager.Companion.PERMISSION_KEYS_REQUEST_CODE
-import org.covidwatch.android.exposurenotification.ExposureNotificationManager.Companion.PERMISSION_START_REQUEST_CODE
 import org.covidwatch.android.extension.send
 import org.covidwatch.android.ui.BaseViewModel
 import org.covidwatch.android.ui.event.Event
 
 open class BaseNotifyOthersViewModel(
-    private val enManager: ExposureNotificationManager,
     positiveDiagnosisRepository: PositiveDiagnosisRepository
 ) : BaseViewModel() {
 
@@ -29,33 +22,7 @@ open class BaseNotifyOthersViewModel(
         }
     }
 
-    fun sharePositiveDiagnosis() {
-        viewModelScope.launch {
-            enManager.isEnabled().success { enabled ->
-                if (enabled) {
-                    shareReport()
-                } else {
-                    withPermission(PERMISSION_START_REQUEST_CODE) {
-                        enManager.start().apply {
-                            success { shareReport() }
-                            failure { handleStatus(it) }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    open fun onTekHistory(teks: MutableList<TemporaryExposureKey>) {
+    open fun sharePositiveDiagnosis() {
         _openVerificationScreen.send()
-    }
-
-    private suspend fun shareReport() {
-        withPermission(PERMISSION_KEYS_REQUEST_CODE) {
-            enManager.temporaryExposureKeyHistory().apply {
-                success { onTekHistory(it) }
-                failure { handleStatus(it) }
-            }
-        }
     }
 }
