+++ app/src/production/java/org/covidwatch/android/di/FlavorSpecificModules.kt
@@ -7,7 +7,6 @@ import org.koin.dsl.module
 val flavorSpecificModule = module {
     viewModel {
         NotifyOthersViewModel(
-            enManager = get(),
             positiveDiagnosisRepository = get()
         )
     }
